tuto: ダークモード機能をstimulusに移行
