+++ app/views/layouts/application.html.erb
@@ -10,7 +10,6 @@
 
     <%= stylesheet_link_tag "application", "data-turbo-track": "reload" %>
     <%= javascript_importmap_tags %>
-    <%= javascript_include_tag "controllers/dark_mode", defer: true %>
       </head>
   <body>
     <%= render 'layouts/header' %>
