+++ app/views/layouts/mailer.html.erb
@@ -6,7 +6,6 @@
       /* Email styles need to be inline */
     </style>
     <%= javascript_importmap_tags %>
-    <%= javascript_include_tag "controllers/dark_mode", defer: true %>
 
   </head>
